chore: fix event handler, styling headless ui tab
